requirements.txt
openai
numpy
scikit-learn
scipy
tenacity
pydantic
tiktoken
python-dotenv
tqdm
